Check models Atdi.DataModels.Sdrns.DeviceServer.Entities

+++ Dev/DataModels/Atdi.DataModels.Sdrn.DeviceServer.Entities/DataLayer/SqlServer/Tables/dbo.SENSOR_ANTENNA.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[SENSOR_ANTENNA](
-	[ID] [bigint] IDENTITY(1,1) NOT NULL,
-	[SENSORID] [bigint] NULL,
+	[ID] [int] IDENTITY(1,1) NOT NULL,
+	[SENSORID] [int] NULL,
 	[CODE] [varchar](50) NULL,
 	[SLEWANG] [float] NULL,
 	[MANUFACTURER] [varchar](50) NULL,
