+++ Dev/DataModels/Atdi.DataModels.Sdrn.DeviceServer.Entities/Entities/ISensorPolygon.cs
@@ -9,13 +9,13 @@ namespace Atdi.DataModels.Sdrns.DeviceServer.Entities
 {
     public interface ISensorPolygon_PK
     {
-        long? Id { get; set; }
+        int? Id { get; set; }
     }
 
     [Entity]
     public interface ISensorPolygon : ISensorPolygon_PK
     {
-        long? SensorId { get; set; }
+        int? SensorId { get; set; }
         double? Lon { get; set; }
         double? Lat { get; set; }
         ISensor SENSOR { get; set; }
